This branch contains development code for a function that will check the distance between two nodes, and if the distance value is negative would push out the nodes to enable a distance between them.
Currently using merge sort does not seem like the correct type of function to use to do the comparisons, as it misses out on some combination of nodes.
The distance checker uses base geometrics,
Cases:
One of the objects compared is a city or a node-> both cannot be as those node types are static and will not overlap
- The object that is not root or city moves
Both object are not city or root
- The objects move in opposite directions by half the amount
The code in this branch is node meant to go to live website, just as a sandbox for the no overlap code.