Improvements to a now complete non optimized overlapping nodes adjuster.
The merge sort was removed as it was not logically right for the the comparison,
a nested for each loop that compares all elements in the array to each other excluding the situation of the same node being compared to itself has been added.
The distance function - needs to be renamed, has been improved to have a minimum distance of 10, and still takes into account the fact that the static nodes should not move if being overlapped with if there is an overlap it increments an overlap counter.
Math formulas corrected.
A counter has been added that is used for a while loop, that runs the clear overlap function until the counter is 0 meaning that all
overlaps are cleared.
Move Node function now has an extra property of type, which updates the origin x and y positions of each node that has been "optimized" and if the node is being "cluster"-ed the radius is not reset to its original size.