Style updates, fix to the text appended to the selected nodes,
Move to front function for text added to bring the text  to front function -> which will raise the text above the the nodes for better visibility.
Functionality to remove circular referencing upon data processing is now part of the data flow, and it removed the infinite loop problem we had when gathering the scene's for the viewer
Reverted version of the opening of the Viewer to a window.open
Right click prevention was re-added.
Auto walk functionality was added, menu is currently under construction
Passing the auto walk functionality to be Added to the memoir